민병호 - khakizone_main 및 KhakiZoneDTO 수정

+++ khaki/src/main/java/co/kr/khaki/zone/KhakiZoneDTO.java
@@ -4,20 +4,20 @@ public class KhakiZoneDTO {
 
 	private int zone_num;			//카키존 넘버
 	private String zone_name;		//카키존 이름
-	private double zone_location_x;	//카키존 위도
-	private double zone_location_y;	//카키존 경도
+	private float zone_location_x;	//카키존 위도
+	private float zone_location_y;	//카키존 경도
 	private String zone_comment;	//카키존 세부설명
 	
-	public double getZone_location_x() {
+	public float getZone_location_x() {
 		return zone_location_x;
 	}
-	public void setZone_location_x(double zone_location_x) {
+	public void setZone_location_x(float zone_location_x) {
 		this.zone_location_x = zone_location_x;
 	}
-	public double getZone_location_y() {
+	public float getZone_location_y() {
 		return zone_location_y;
 	}
-	public void setZone_location_y(double zone_location_y) {
+	public void setZone_location_y(float zone_location_y) {
 		this.zone_location_y = zone_location_y;
 	}
 	public int getZone_num() {
